Refactor v2 collection evaluation helpers
